Add files via upload
Changed images single-box-step3 and step-5 to remove heading text and to scale down size. Added an image for the was-console.
Edited single-box.md to add steps that accurately showed the AWS configuration process. Made some other edits for clarity and simplicity of language. Tried to make each step a distinct instruction, that is, took explanatory text out of the step instructions.